alter lesson number to be integer, get all words in lesson one to table

+++ ilocano-backend/wordbankdb/models.py
@@ -9,7 +9,7 @@ class WordBank(models.Model):
     english = models.CharField(max_length=30)
     category = models.CharField(max_length=20 , null=True, blank=True)
     variations = ArrayField(models.CharField(max_length=20), null=True, blank=True)
-    lesson_number = models.DecimalField(max_digits=3, decimal_places=1, null=True, blank=True)
+    lesson_number = models.IntegerField(null=True, blank=True)
 
     def __str__(self) -> str:
         return f"{self.ilokano} {self.english}"
